Moved flaps handle to common file
