+++ fbw-common/src/wasm/systems/systems/src/hydraulic/command_sensor_unit.rs
@@ -1,0 +1,35 @@
+use crate::simulation::{
+    InitContext, Read, SimulationElement, SimulatorReader, VariableIdentifier,
+};
+
+/// A struct to read the handle position
+pub struct FlapsHandle {
+    handle_position_id: VariableIdentifier,
+    position: u8,
+    previous_position: u8,
+}
+
+impl FlapsHandle {
+    pub fn new(context: &mut InitContext) -> Self {
+        Self {
+            handle_position_id: context.get_identifier("FLAPS_HANDLE_INDEX".to_owned()),
+            position: 0,
+            previous_position: 0,
+        }
+    }
+
+    pub fn position(&self) -> u8 {
+        self.position
+    }
+
+    pub fn previous_position(&self) -> u8 {
+        self.previous_position
+    }
+}
+
+impl SimulationElement for FlapsHandle {
+    fn read(&mut self, reader: &mut SimulatorReader) {
+        self.previous_position = self.position;
+        self.position = reader.read(&self.handle_position_id);
+    }
+}
