+++ fbw-common/src/wasm/systems/systems/src/hydraulic/mod.rs
@@ -37,6 +37,7 @@ pub mod brake;
 pub mod brake_circuit;
 pub mod bypass_pin;
 pub mod cargo_doors;
+pub mod command_sensor_unit;
 pub mod electrical_generator;
 pub mod electrical_pump_physics;
 pub mod flap_slat;
